requirements.txt
asgiref==3.4.1
Django==3.2.13
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
idna==3.3
Pillow==8.4.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
